Init refactoring
