+++ zds/forum/views.py
@@ -12,9 +12,9 @@ import json
 
 from forms import TopicForm, PostForm
 from models import Category, Forum, Topic, Post, \
-    PostDislike, PostLike, follow, never_read, mark_read
+    follow, never_read, mark_read
 from zds.utils import render_template, slugify
-from zds.utils.models import Alert
+from zds.utils.models import Alert, CommentLike, CommentDislike
 from zds.utils.paginator import paginator_range
 from zds.member.models import Profile
 
@@ -502,19 +502,19 @@ def like_post(request):
     
     if post.author.pk != request.user.pk:
         # Making sure the user is allowed to do that
-        if PostLike.objects.filter(user__pk=user.pk, posts__pk=post_pk).count()==0:
-            like=PostLike()
+        if CommentLike.objects.filter(user__pk=user.pk, comments__pk=post_pk).count()==0:
+            like=CommentLike()
             like.user=user
             like.posts=post
             post.like=post.like+1
             post.save()
             like.save()
-            if PostDislike.objects.filter(user__pk=user.pk, posts__pk=post_pk).count()>0:
-                PostDislike.objects.filter(user__pk=user.pk, posts__pk=post_pk).all().delete()
+            if CommentDislike.objects.filter(user__pk=user.pk, comments__pk=post_pk).count()>0:
+                CommentDislike.objects.filter(user__pk=user.pk, comments__pk=post_pk).all().delete()
                 post.dislike=post.dislike-1
                 post.save()
         else:
-            PostLike.objects.filter(user__pk=user.pk, posts__pk=post_pk).all().delete()
+            CommentLike.objects.filter(user__pk=user.pk, comments__pk=post_pk).all().delete()
             post.like=post.like-1
             post.save()
 
@@ -539,19 +539,19 @@ def dislike_post(request):
 
     if post.author.pk != request.user.pk:
         # Making sure the user is allowed to do that
-        if PostDislike.objects.filter(user__pk=user.pk, posts__pk=post_pk).count()==0:
-            dislike=PostDislike()
+        if CommentDislike.objects.filter(user__pk=user.pk, comments__pk=post_pk).count()==0:
+            dislike=CommentDislike()
             dislike.user=user
             dislike.posts=post
             post.dislike=post.dislike+1
             post.save()
             dislike.save()
-            if PostLike.objects.filter(user__pk=user.pk, posts__pk=post_pk).count()>0:
-                PostLike.objects.filter(user__pk=user.pk, posts__pk=post_pk).all().delete()
+            if CommentLike.objects.filter(user__pk=user.pk, comments__pk=post_pk).count()>0:
+                CommentLike.objects.filter(user__pk=user.pk, comments__pk=post_pk).all().delete()
                 post.like=post.like-1
                 post.save()
         else :
-            PostDislike.objects.filter(user__pk=user.pk, posts__pk=post_pk).all().delete()
+            CommentDislike.objects.filter(user__pk=user.pk, comments__pk=post_pk).all().delete()
             post.dislike=post.dislike-1
             post.save()
